+++ src/preload/index.d.ts
@@ -23,14 +23,17 @@ import type {
   OS,
   ModelInfo,
   TodoItem,
+  UsageDataRow,
+  EnvironmentVariable,
 } from '@common/types';
 
 export interface ApplicationAPI {
+  openLogsDirectory: () => Promise<boolean>;
   loadSettings: () => Promise<SettingsData>;
   saveSettings: (settings: SettingsData) => Promise<SettingsData>;
   startProject: (baseDir: string) => void;
   stopProject: (baseDir: string) => void;
-  restartProject: (baseDir: string) => void;
+  restartProject: (baseDir: string, startupMode?: StartupMode) => void;
   runPrompt: (baseDir: string, prompt: string, mode?: Mode) => void;
   redoLastUserPrompt: (baseDir: string, mode: Mode, updatedPrompt?: string) => void;
   answerQuestion: (baseDir: string, answer: string) => void;
@@ -92,6 +95,8 @@ export interface ApplicationAPI {
   clearReleaseNotes: () => Promise<void>;
   getOS: () => Promise<OS>;
   loadModelsInfo: () => Promise<Record<string, ModelInfo>>;
+  queryUsageData: (from: string, to: string) => Promise<UsageDataRow[]>;
+  getEffectiveEnvironmentVariable: (key: string, baseDir?: string) => Promise<EnvironmentVariable | undefined>;
 
   addResponseChunkListener: (baseDir: string, callback: (event: Electron.IpcRendererEvent, data: ResponseChunkData) => void) => string;
   removeResponseChunkListener: (listenerId: string) => void;